feat(ONB-204): add composable for keyword density analysis and integrate into UI
